refactor: rename navbar back to scaffold_with_navbar
