+++ lib/router.dart
@@ -2,7 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
 import 'package:messless/screens/home.dart';
 import 'package:messless/screens/login.dart';
-import 'package:messless/widgets/navbar.dart';
+import 'package:messless/widgets/scaffold_with_navbar.dart';
 
 enum RouterDestinations {
   home(url: '/'),
