+++ lib/widgets/scaffold_with_navbar.dart
